created a event module, practice with dependency injection, added config, logging, exceptions

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, HttpCode, Param, ParseIntPipe, Patch, Post, ValidationPipe } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, Logger, NotFoundException, Param, ParseIntPipe, Patch, Post, ValidationPipe } from "@nestjs/common";
 import { CreateEventDto } from "./create-event.dto";
 import { UpdateEventDto } from "./update-event.dto";
 import { Event } from "./event.entity";
@@ -7,6 +7,7 @@ import { InjectRepository } from "@nestjs/typeorm";
 
 @Controller('/events')
 export class EventsController{
+    private readonly logger = new Logger(EventsController.name);
 
     constructor(
         @InjectRepository(Event)
@@ -15,17 +16,30 @@ export class EventsController{
 
     @Get()
     async findAll(){
-        return await this.repository.find();
+        this.logger.log(`GET: FIND ALL ROUTE`);
+
+        const events = await this.repository.find();
+
+        this.logger.debug(`Found ${events.length} events`);
+
+        return events;
     }
 
     @Get(':id')
     async findOne(@Param('id', ParseIntPipe) id: number){
-        console.log(typeof id);
-        return await this.repository.findOneBy({id: id});
+        this.logger.log(`GET: FIND ONE ROUTE`);
+         
+        const event = await this.repository.findOneBy({id: id});
+
+        if(!event)
+            throw new NotFoundException();
+
+        return event;
     }
 
     @Post()
     async create(@Body() input: CreateEventDto){
+        this.logger.log(`POST: CREATE ROUTE`);
         return await this.repository.save({
             ...input,
             when: new Date(input.when)
@@ -34,7 +48,11 @@ export class EventsController{
 
     @Patch(':id')
     async update(@Param('id') id, @Body() input: UpdateEventDto){
+        this.logger.log(`PATCH: UPDATE ROUTE`);
         const event = await this.repository.findOneBy({id: id})
+
+        if(!event)
+            throw new NotFoundException();
 
         return await this.repository.save({
             ...event,
@@ -46,7 +64,12 @@ export class EventsController{
     @Delete(':id')
     @HttpCode(204)
     async remove(@Param('id') id){
+        this.logger.log(`DELETE: REMOVE ROUTE`);
         const event = await this.repository.findOneBy({id: id});
+
+        if(!event)
+            throw new NotFoundException();
+
         await this.repository.remove(event);
     }
 }
